fixed rounding issue my tests exposed

+++ src/src/com/rocktail/hobbitutilst/models/TroopsTrainingCalculationResult.java
@@ -3,15 +3,15 @@ package com.rocktail.hobbitutilst.models;
 import java.util.Observable;
 
 public class TroopsTrainingCalculationResult extends Observable {
-	private int _footTroopsAmount;
-	private int _rangedTroopsAmount;
-	private int _mountedTroopsAmount;
+	private long _footTroopsAmount;
+	private long _rangedTroopsAmount;
+	private long _mountedTroopsAmount;
 	
 	/**
 	 * Gets calculated amount of foot troops to produce
 	 * @return
 	 */
-	public int getFootTroopsAmount() {
+	public long getFootTroopsAmount() {
 		return _footTroopsAmount;
 	}
 	
@@ -19,7 +19,7 @@ public class TroopsTrainingCalculationResult extends Observable {
 	 * Sets amount of foot troops
 	 * @param _footTroopsAmount
 	 */
-	public void setFootTroopsAmount(int _footTroopsAmount) {
+	public void setFootTroopsAmount(long _footTroopsAmount) {
 		this._footTroopsAmount = _footTroopsAmount;
 		setChanged();
 		notifyObservers();
@@ -29,7 +29,7 @@ public class TroopsTrainingCalculationResult extends Observable {
 	 * Gets calculated amount of ranged troops to produce
 	 * @return
 	 */
-	public int getRangedTroopsAmount() {
+	public long getRangedTroopsAmount() {
 		return _rangedTroopsAmount;
 	}
 	
@@ -37,7 +37,7 @@ public class TroopsTrainingCalculationResult extends Observable {
 	 * Sets amount of ranged troops
 	 * @param _rangedTroopsAmount
 	 */
-	public void setRangedTroopsAmount(int _rangedTroopsAmount) {
+	public void setRangedTroopsAmount(long _rangedTroopsAmount) {
 		this._rangedTroopsAmount = _rangedTroopsAmount;
 		setChanged();
 		notifyObservers();
@@ -47,7 +47,7 @@ public class TroopsTrainingCalculationResult extends Observable {
 	 * Gets calculated amount of mounted troops to produce
 	 * @return
 	 */
-	public int getMountedTroopsAmount() {
+	public long getMountedTroopsAmount() {
 		return _mountedTroopsAmount;
 	}
 	
@@ -55,7 +55,7 @@ public class TroopsTrainingCalculationResult extends Observable {
 	 * Sets amount of mounted troops
 	 * @param _mountedTroopsAmount
 	 */
-	public void setMountedTroopsAmount(int _mountedTroopsAmount) {
+	public void setMountedTroopsAmount(long _mountedTroopsAmount) {
 		this._mountedTroopsAmount = _mountedTroopsAmount;
 		setChanged();
 		notifyObservers();
